we makin it out of the node2d with this one (godot rust hello world crap)
