+++ pets-gd/scenes/battle_engine.tscn
@@ -1,0 +1,3 @@
+[gd_scene format=3 uid="uid://kgl5788x405y"]
+
+[node name="BattleEngine" type="BattleEngine"]
