+++ pets-lib/src/dialogue.rs
@@ -1,0 +1,10 @@
+type DialogueOptions = Vec<DialogueNode>;
+
+struct DialogueNode {
+    text: String,
+    options: DialogueOptions,
+}
+
+fn dialogue_yes_no() {
+    unimplemented!();
+}
